+++ test/geolix/adapter/mmdb2/loader_test.exs
@@ -43,6 +43,7 @@ defmodule Geolix.Adapter.MMDB2.Database.LoaderTest do
     assert %Result.City{} = Geolix.lookup("2.125.160.216", where: :system_env)
 
     System.delete_env(var)
+    Geolix.unload_database(:system_env)
   end
 
   test "system environment configuration (default value)" do
@@ -54,6 +55,8 @@ defmodule Geolix.Adapter.MMDB2.Database.LoaderTest do
 
     assert :ok = Geolix.load_database(db)
     assert %Result.City{} = Geolix.lookup("2.125.160.216", where: :system_env_default)
+
+    Geolix.unload_database(:system_env_default)
   end